Add 2 samples for express.js.
The sample "mongo_api_test" is used to show how to use MongoDB API for Node.js.

The sample "fb_test1" is not used to show the functions of express.js.
It is used to show how to use Facebook API, which just uses Node.js as its web server.